dev/assistantfoodfitness/requirements.txt
openai
mysql-connector-python
pyyaml
